+++ src/main/kotlin/com/example/domain/login/usecase/AddNewUserUseCase.kt
@@ -1,8 +1,8 @@
 package com.example.domain.login.usecase
 
-import com.example.domain.login.adapter.UserRepository
 import com.example.domain.login.model.UserDomainModel
 import com.example.domain.login.model.toData
+import com.example.domain.login.source.UserRepository
 
 class AddNewUserUseCase(
     private val userRepository: UserRepository
