+++ src/main/kotlin/com/example/infrastructure/config/Application.kt
@@ -4,6 +4,7 @@ import com.example.di.AppComponent
 import com.example.di.DaggerAppComponent
 import com.example.infrastructure.config.plugins.configureRouting
 import com.example.infrastructure.config.plugins.configureSockets
+import io.ktor.http.*
 import io.ktor.server.engine.*
 import io.ktor.server.netty.*
 
@@ -13,8 +14,8 @@ val appComponent: AppComponent = DaggerAppComponent.create()
 fun main() {
 
     ServerConfig.defaultConfig()
-    embeddedServer(Netty, port = System.getenv("PORT").toInt()) {
-//    embeddedServer(Netty, port = 8080, host = "0.0.0.0") {
+//    embeddedServer(Netty, port = System.getenv("PORT")?.toInt() ?: DEFAULT_PORT) {
+    embeddedServer(Netty, port = 8080, host = "0.0.0.0") {
         configureSockets()
         configureRouting()
     }.start(wait = true)
